move normalize_data_format to seperate file
